Update printTable method to print an extra space after a comma in list
