+++ test/fixtures/output/network/describe-help.golden
@@ -9,6 +9,7 @@ Describe Confluent network "n-abcde1".
   $ confluent network describe n-abcde1
 
 Flags:
+      --context string       CLI context name.
       --environment string   Environment ID.
   -o, --output string        Specify the output format as "human", "json", or "yaml". (default "human")
