+++ test/fixtures/output/network/describe.golden
@@ -1,12 +1,12 @@
-+-------------------------+----------------------------+
-| ID                      | n-abcde1                   |
-| Environment ID          | env-00000                  |
-| Name                    | prod-aws-us-east1          |
-| Cloud                   | AWS                        |
-| Region                  | us-east-1                  |
-| CIDR                    | 10.200.0.0/16              |
-| Zones                   | use1-az1,use1-az2,use1-az3 |
-| DNS Resolution          | CHASED_PRIVATE             |
-| Phase                   | READY                      |
-| Active Connection Types | PRIVATELINK,TRANSITGATEWAY |
-+-------------------------+----------------------------+
++-------------------------+------------------------------+
+| ID                      | n-abcde1                     |
+| Environment ID          | env-00000                    |
+| Name                    | prod-aws-us-east1            |
+| Cloud                   | AWS                          |
+| Region                  | us-east-1                    |
+| CIDR                    | 10.200.0.0/16                |
+| Zones                   | use1-az1, use1-az2, use1-az3 |
+| DNS Resolution          | CHASED_PRIVATE               |
+| Phase                   | READY                        |
+| Active Connection Types | PRIVATELINK, TRANSITGATEWAY  |
++-------------------------+------------------------------+
